+++ src/const.js
@@ -2,26 +2,7 @@ const SORTER_TYPES = ['day', 'event', 'time', 'price', 'offer'];
 const FILTER_TYPES = ['everything', 'future', 'present', 'past'];
 const POINT_TYPES = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
 
-const MOCK_PHOTOGEN_MIN = 0;
-const MOCK_PHOTOGEN_MAX = 100;
-
-const MOCK_DESCGEN_MIN = 0;
-const MOCK_DESCGEN_MAX = 5;
-
-const MOCK_LOREM = [
-  'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-  'Cras aliquet varius magna, non porta ligula feugiat eget.',
-  'Fusce tristique felis at fermentum pharetra.',
-  'Aliquam id orci ut lectus varius viverra.',
-  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
-  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.',
-  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.',
-  'Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat.',
-  'Nunc fermentum tortor ac porta dapibus.',
-  'In rutrum ac purus sit amet tempus.',
-];
-
 const EVENT_COUNT = 6;
 const EVENT_VIEW_AMOUNT = 5;
 
-export {SORTER_TYPES, FILTER_TYPES, POINT_TYPES, MOCK_PHOTOGEN_MIN, MOCK_PHOTOGEN_MAX, MOCK_LOREM, MOCK_DESCGEN_MIN, MOCK_DESCGEN_MAX, EVENT_COUNT, EVENT_VIEW_AMOUNT};
+export {SORTER_TYPES, FILTER_TYPES, POINT_TYPES, EVENT_COUNT, EVENT_VIEW_AMOUNT};
